+++ ngx-fudis/projects/dev/src/app/app.component.ts
@@ -10,7 +10,7 @@ import {
 	FudisGridService,
 	FudisTranslationService,
 	FudisBreakpointService,
-	FudisReloadErrorsErrorSummaryService,
+	FudisErrorSummaryService,
 } from 'ngx-fudis';
 import { DOCUMENT } from '@angular/common';
 
@@ -31,7 +31,7 @@ export class AppComponent implements OnInit {
 		private _gridService: FudisGridService,
 		private _fudisLanguage: FudisTranslationService,
 		private _alertService: FudisAlertService,
-		private _reloadErrorsErrorSummaryService: FudisReloadErrorsErrorSummaryService,
+		private _errorSummaryService: FudisErrorSummaryService,
 
 		private _breakpointService: FudisBreakpointService
 	) {
@@ -145,7 +145,7 @@ export class AppComponent implements OnInit {
 			this._fudisLanguage.setLanguage('en');
 		}
 
-		this._reloadErrorsErrorSummaryService.reloadErrors();
+		this._errorSummaryService.reloadErrors();
 	}
 
 	openDialog(): void {
