+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/error-summary/readme.mdx
@@ -9,4 +9,4 @@ import * as ErrorSummaryStories from './error-summary.stories.ts';
 
 ### Related services
 
-[Reload errors](/docs/services-reload-errors--documentation)
+[Reload errors](/docs/services-error-summary--documentation)
