+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/form/form.stories.ts
@@ -8,7 +8,7 @@ import { RouterModule } from '@angular/router';
 import { BehaviorSubject } from 'rxjs';
 import { FudisDropdownOption, FudisRadioButtonOption, FudisFormErrors, FudisDateRangeItem } from '../../../types/forms';
 
-import { FudisReloadErrorsErrorSummaryService } from '../../../services/form/error-summary/reload-errors/reload-errors-error-summary.service';
+import { FudisErrorSummaryService } from '../../../services/form/error-summary/error-summary.service';
 import { FudisGroupValidators, FudisValidators } from '../../../utilities/form/validators';
 import { FormComponent } from './form.component';
 import { FudisTranslationService } from '../../../services/translation/translation.service';
@@ -151,7 +151,7 @@ import { FudisFocusService } from '../../../services/focus/focus.service';
 })
 class FormContentExampleComponent implements OnInit {
 	constructor(
-		private _reloadErrorsErrorSummaryService: FudisReloadErrorsErrorSummaryService,
+		private _errorSummaryService: FudisErrorSummaryService,
 		private _translationService: FudisTranslationService,
 		private _focusService: FudisFocusService
 	) {}
@@ -264,7 +264,7 @@ class FormContentExampleComponent implements OnInit {
 		if (this.formExample.invalid) {
 			this._closed = false;
 			this.errorSummaryVisible = true;
-			this._reloadErrorsErrorSummaryService.reloadErrors();
+			this._errorSummaryService.reloadErrors();
 		} else {
 			this.errorSummaryVisible = false;
 		}
