+++ ngx-fudis/projects/ngx-fudis/src/lib/ngx-fudis.module.ts
@@ -52,8 +52,8 @@ import { DropdownMenuItemComponent } from './components/dropdown-menu/dropdown-m
 import { FudisDropdownMenuItemService } from './components/dropdown-menu/dropdown-menu-item/dropdown-menu-item.service';
 import { FudisFocusService } from './services/focus/focus.service';
 import { ErrorMessageComponent } from './components/form/error-message/error-message.component';
+import { FudisInternalErrorSummaryService } from './services/form/error-summary/internal-error-summary.service';
 import { FudisErrorSummaryService } from './services/form/error-summary/error-summary.service';
-import { FudisReloadErrorsErrorSummaryService } from './services/form/error-summary/reload-errors/reload-errors-error-summary.service';
 import { ErrorSummaryComponent } from './components/form/error-summary/error-summary.component';
 import { ExpandableComponent } from './components/expandable/expandable.component';
 
@@ -257,12 +257,12 @@ import { FudisBreakpointService } from './services/breakpoint/breakpoint.service
 		FudisBreakpointService,
 		FudisDialogService,
 		FudisDropdownMenuItemService,
-		FudisErrorSummaryService,
+		FudisInternalErrorSummaryService,
 		FudisFocusService,
 		FudisGridService,
 		FudisIdService,
 		FudisLanguageBadgeGroupService,
-		FudisReloadErrorsErrorSummaryService,
+		FudisErrorSummaryService,
 		FudisTranslationService,
 	],
 })
