+++ ngx-fudis/projects/ngx-fudis/src/lib/services/form/error-summary/readme.stories.mdx
@@ -1,0 +1,35 @@
+import { Meta } from '@storybook/blocks';
+
+<Meta title="Services/Error Summary" />
+
+# Error Summary
+
+## Functions
+
+### Reload errors
+
+Reload errors is an error summary service functionality that updates the visible list of errors shown in the Fudis Error Summary component.
+
+Import `FudisErrorSummaryService` and add the parameter to constructor:
+
+```
+import { FudisErrorSummaryService } from 'ngx-fudis';
+
+...
+
+constructor(
+  private _service: FudisErrorSummaryService
+) {}
+```
+
+When calling `reloadErrors` function, you can set an optional timeout delay in milliseconds:
+
+```
+this._service.reloadErrors(); // sets timeout to 0ms
+
+this._service.reloadErrors(100); // sets timeout to 100ms
+```
+
+## Related components
+
+[Error-summary](/docs/components-form-error-summary--documentation)
