--- ngx-fudis/projects/ngx-fudis/src/lib/services/form/error-summary/reload-errors/readme.stories.mdx
@@ -1,33 +1,0 @@
-import { Meta } from '@storybook/blocks';
-
-<Meta title="Services/Reload Errors" />
-
-# Reload errors
-
-Reload errors is an error summary service functionality that updates the visible list of errors shown in the Fudis Error Summary component.
-
-## Usage Guidelines
-
-Import `FudisReloadErrorsErrorSummaryService` and add the parameter to constructor.
-
-```
-import { FudisReloadErrorsErrorSummaryService } from 'ngx-fudis';
-
-...
-
-constructor(
-  private _service: FudisReloadErrorsErrorSummaryService
-) {}
-```
-
-When calling `reloadErrors` function, you can set an optional timeout delay in milliseconds.
-
-```
-this._service.reloadErrors(); // sets timeout to 0ms
-
-this._service.reloadErrors(100); // sets timeout to 100ms
-```
-
-### Related components
-
-[Error-summary](/docs/components-form-error-summary--documentation)
